review 리팩토링
