+++ routes/db.js
@@ -1,21 +1,21 @@
 const router = require('express').Router();
-const reviewController = require('../controllers/review');
+const mongodb = require('../controllers/mongodb');
 
 router.get('/', async (req, res) => {
-    res.json({result: true});
-})
+  res.json({ result: true });
+});
 
 router.post('/insert', async (req, res) => {
-    console.log(req);
-    const result = await reviewController.insertUserReview(req.body);
+  console.log(req);
+  const result = await mongodb.insertUserReview(req.body);
 
-    res.json(result);
-})
+  res.json(result);
+});
 
 router.post('/find', async (req, res) => {
-    const review = await reviewController.getReviews(req.body);
+  const review = await mongodb.getReviews(req.body);
 
-    res.json(review);
-})
+  res.json(review);
+});
 
 module.exports = router;
