+++ routes/request.js
@@ -1,9 +1,9 @@
 const router = require('express').Router();
-const libKakaoWork = require('../libs/kakaoWork');
 const keywordModal = require('../msgGenerator/keywordModal.msg');
 const reviewWriteModal = require('../msgGenerator/review/reviewWriteModal.msg');
 const reviewSearchModal = require('../msgGenerator/review/reviewSearchModal.msg');
 const accountModal = require('../msgGenerator/accountModal.msg');
+
 /**
  *  @author  dongjin
  *  @brief
@@ -30,15 +30,14 @@ router.post('/', async (req, res, next) => {
 
     case 'review_request': {
       // "한줄평 쓰러가기" 버튼 눌렀을 때 모달 전송
-      //value: yesterdaydata+userid+action_name
       const msg = reviewWriteModal(value_json);
-      return res.json(msg); //사용자에게 모달 띄움
+      return res.json(msg);
     }
 
     case 'review_search': {
       // "멘토링 한줄평 검색" 버튼 눌렀을 때 모달 전송
       const msg = reviewSearchModal();
-      return res.json(msg); //사용자에게 모달 띄움
+      return res.json(msg);
     }
 
     case 'montoring_setting': {
